revert 6251644, and partially 1bd9a35
in 6251644 the change to the makeflags sed also made it less specific, acting on lines matching "MAKEFLAGS" instead of "^#MAKEFLAGS". the change to the pacman color sed in 1bd9a35 once given the same correction only saves 1 character over the previous commit's more readable expression where "Color" appears twice, so it's reverted to that.